requirements.txt
langchain==0.2.14
langchain-core==0.2.34
langchain-openai==0.1.19
langchain-community==0.2.12
transformers
aiolimiter
pyarrow
pandas
numpy
azure-storage-blob==12.22.0
python-dotenv
pyyaml
boto3
botocore
tenacity
openpyxl
python-dateutil
python-docx
streamlit>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
pyarrow>=12.0.1
openpyxl>=3.1.2
